feat : adjust the trayer cmd

+++ scripts/trayertoggle.sh
@@ -1,3 +1,10 @@
-#!/bin/bash
-killall trayer
-trayer --edge top --align right --width 10 --transparent true --alpha 127 --expand true --height 18 --tint 0xffff00ff
+#!/bin/sh
+
+tra_res=$(ps -ef|grep -vE "grep|trayertoggle" |grep -i trayer)
+echo $tra_res
+if [ "" = "$tra_res" ]; then
+    trayer --edge bottom --align right --width 10 --transparent true --alpha 250 --expand true --height 20 --tint 0xffff00ff &
+else
+    killall trayer
+fi
+
